Move scenario build methods.
