+++ examples/egoless.py
@@ -3,8 +3,8 @@ from pathlib import Path
 import gym
 from tools.argument_parser import default_argument_parser
 
-from cli.studio import build_scenarios
 from smarts.core.utils.episodes import episodes
+from smarts.sstudio.scenario_construction import build_scenarios
 
 
 def main(scenarios, headless, num_episodes, max_episode_steps=None):
